+++ CHANGELOG.md
@@ -11,6 +11,9 @@
 - added Show task info button for quick access to task information
 - image messages are sent also when Include Context Files is disabled
 - properly initializing internal task in project
+- added Updated Files section with information about the changes
+- BMAD mode consecutive prompts are now properly sending when working via browser
+- sync task baseDir with project baseDir to handle migrated tasks
 
 ## [0.52.0]